+++ recipe.lua
@@ -25,7 +25,7 @@ return {
                 execute = "/usr/local/bin/script.sh add-missing-indices",
             },
         },
-        -- Raw options.
+        -- Pod options.
         options = {
             "--userns=host",
             "--tty",
@@ -70,7 +70,7 @@ return {
             restart = "on-failure",
             image = "simple3:1.1",
             options = {
-                "--env ADMIN_TOKEN='1234567890'",
+                "--env=ADMIN_TOKEN=1234567890",
             },
         },
     },