+++ src/container.lua
@@ -91,7 +91,9 @@ global function container__create(container, pod, simulate)
     -- container options
     -- if not supported by pods, add them directly to the podman run command
     if not table.is_nil_or_empty(container.options) then
-        commands[#commands + 1] = table.concat(container.options, " ")
+        for i = 1, #container.options do
+            commands[#commands + 1] = string.escape_shell(container.options[i])
+        end
     end
 
     -- container image
@@ -101,7 +103,9 @@ global function container__create(container, pod, simulate)
     -- commands
     -- see: podman run --detach image:tag command
     if not table.is_nil_or_empty(container.commands) then
-        commands[#commands + 1] = table.concat(container.commands, " ")
+        for i = 1, #container.commands do
+            commands[#commands + 1] = string.escape_shell(container.commands[i])
+        end
     end
 
     -- create and execute final podman command